# Arquivos/Swagger doc.yaml
openapi: 3.0.1
info:
  title: E-Commerce Swagger
  description: API desenvolvida para sistema de E-Commerce
  version: '1'
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /itens-carrinho:
    get:
      tags:
        - Carrinho
      summary: Exibe os itens do carrinho
      operationId: verCarrinho
      responses:
        '200':
          description: Exibição dos itens realizado com sucesso
          content:
            '*/*':
              example:
                - produto:
                    id: f9c473bf-afb8-4eb2-9d32-d1c2b9498408
                    nome: Blusa B
                    quantidadeEstoque: 326
                    preco: 87.5
                    links:
                      - rel: self
                        href: http://localhost:8080/produtos/id/f9c473bf-afb8-4eb2-9d32-d1c2b9498408
                  quantidade: 2
                  valorTotalProduto: 175
                - produto:
                    id: b7545abe-eec2-4537-88fc-63667782789c
                    nome: Meia B
                    quantidadeEstoque: 56
                    preco: 6.44
                    links:
                      - rel: self
                        href: http://localhost:8080/produtos/id/b7545abe-eec2-4537-88fc-63667782789c
                  quantidade: 7
                  valorTotalProduto: 45.08
                - produto:
                    id: 8fcfe445-864c-4d96-b8c0-7d75acb45dc5
                    nome: Tênis B
                    quantidadeEstoque: 91
                    preco: 6.44
                    links:
                      - rel: self
                        href: http://localhost:8080/produtos/id/8fcfe445-864c-4d96-b8c0-7d75acb45dc5
                  quantidade: 4
                  valorTotalProduto: 25.76
                - Quantidade de itens: 13
                  Quantidade de produtos: 3
                  Valor total do carrinho: 245.84
        '400':
          description: Carrinho inexistente.
          content:
            '*/*':
              example: Usuário Bernardo Kauê da Silva não possui carrinho ativo.
        '401':
          description: Não autorizado
        '403':
          description: Acesso proibido
      security:
        - ecommerce: []
    put:
      tags:
        - Carrinho
      summary: Insere ou altera um item no carrinho
      operationId: inserirProduto
      parameters:
        - name: id_produto
          in: query
          required: true
          content:
            '*/*':
              example: 8fcfe445-864c-4d96-b8c0-7d75acb45dc5
        - name: quantidade
          in: query
          required: true
          content:
            '*/*':
              example: 4
      responses:
        '201':
          description: Inserção do item realizada com sucesso
          content:
            '*/*':
              example:
                - produto:
                    id: 8fcfe445-864c-4d96-b8c0-7d75acb45dc5
                    nome: Tênis B
                    quantidadeEstoque: 87
                    preco: 6.44
                    links:
                      - rel: self
                        href: http://localhost:8080/produtos/id/8fcfe445-864c-4d96-b8c0-7d75acb45dc5
                  quantidade: 3
                  valorTotalProduto: 19.32
                - Quantidade de itens: 3
                  Quantidade de produtos: 1
                  Valor total do carrinho: 19.32
        '400':
          description: Id inválida.
        '401':
          description: Não autorizado
        '403':
          description: Acesso proibido
        '404':
          description: Produto não encontrado
      security:
        - ecommerce: []
    delete:
      tags:
        - Carrinho
      summary: Remove todos os itens do carrinho
      operationId: cancelaCarrinho
      responses:
        '200':
          description: Carrinho cancelado.
        '400':
          description: Carrinho inexistente.
          content:
            '*/*':
              example: Usuário Bernardo Kauê da Silva não possui carrinho ativo.
        '401':
          description: Não autorizado
        '403':
          description: Acesso proibido
      security:
        - ecommerce: []
...
  /itens-carrinho/venda:
    post:
      tags:
        - Carrinho
      summary: Efetiva a venda ao usuário
      operationId: confirmaVenda
      responses:
        '201':
          description: Efetivação da venda realizada com sucesso
          content:
            '*/*':
              example: Venda confirmada.
        '400':
          description: Carrinho inexistente.
          content:
            '*/*':
              example: Usuário Bernardo Kauê da Silva não possui carrinho ativo.
        '401':
          description: Não autorizado
        '403':
          description: Acesso proibido
        '409':
          description: Quantidade do produto no carrinho tornou-se maior que a disponível no estoque entre a inserção no carrinho e a efetivação da venda.
          content:
            '*/*':
              example: Quantidade de Tênis B indisponível no estoque.
      security:
        - ecommerce: []
...
  /produtos/id/{id_produto}:
    get:
      tags:
        - Produtos
      summary: Busca um produto por ID
      operationId: buscarProdutoPorId
      parameters:
        - name: id_produto
          in: path
          required: true
          content:
            '*/*':
              example: f9c473bf-afb8-4eb2-9d32-d1c2b9498408
      responses:
        '200':
          description: Busca de produto realizado com sucesso
          content:
            '*/*':
              example:
                id: f9c473bf-afb8-4eb2-9d32-d1c2b9498408
                nome: Blusa B
                quantidadeEstoque: 500
                preco: 10
                _links:
                  Lista de Produtos:
                    href: http://localhost:8080/produtos
        '400':
          description: Id inválida
        '404':
          description: Produto não encontrado
...
  /itens-carrinho/{id_produto}:
    delete:
      tags:
        - Carrinho
      summary: Remove um item do carrinho por ID
      operationId: removeItemCarrinho
      parameters:
        - name: id_produto
          in: path
          required: true
          content:
            '*/*':
              example: 8fcfe445-864c-4d96-b8c0-7d75acb45dc5
      responses:
        '200':
          description: Remoção do item realizado com sucesso
          content:
            '*/*':
              example: Item Tênis B removido.
        '400':
          description: Id inválida.
        '401':
          description: Não autorizado
        '403':
          description: Acesso proibido
        '404':
          description: Produto não encontrado
      security:
        - ecommerce: []
components:
...
  securitySchemes:
    ecommerce:
      type: http
      scheme: basic